+++ pyproject.toml
@@ -26,6 +26,7 @@ pydantic = "^1.9"
 python-dotenv = "^0.20.0"
 beartype = "^0.10.4"
 sortedcontainers = "^2.4.0"
+petl = "^1.7.9"
 
 [tool.poetry.dev-dependencies]
 black = "*"
@@ -33,6 +34,7 @@ ipython = "*"
 mypy = "*"
 pip = "*"
 pytest = "*"
+ipykernel = "^6.13.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
